requirements.txt
streamlit
snowflake-snowpark-python
pandas
snowflake.core
snowflake-ml-python
dotenv